Split module `dateformatter` into `api`, `impl` and `test`.
