+++ app/build.gradle.kts
@@ -199,6 +199,7 @@ dependencies {
     allLibraries()
     allFeatures()
     implementation(projects.libraries.matrix.impl)
+    implementation(projects.libraries.dateformatter.impl)
     implementation(projects.tests.uitests)
     implementation(projects.anvilannotations)
     anvil(projects.anvilcodegen)
