+++ features/messages/src/test/kotlin/io/element/android/features/messages/fixtures/timelineItemsFactory.kt
@@ -16,7 +16,6 @@
 
 package io.element.android.features.messages.fixtures
 
-import io.element.android.features.messages.fakes.FakeDaySeparatorFormatter
 import io.element.android.features.messages.timeline.factories.TimelineItemsFactory
 import io.element.android.features.messages.timeline.factories.event.TimelineItemContentFactory
 import io.element.android.features.messages.timeline.factories.event.TimelineItemContentFailedToParseMessageFactory
@@ -31,6 +30,7 @@ import io.element.android.features.messages.timeline.factories.event.TimelineIte
 import io.element.android.features.messages.timeline.factories.event.TimelineItemEventFactory
 import io.element.android.features.messages.timeline.factories.virtual.TimelineItemDaySeparatorFactory
 import io.element.android.features.messages.timeline.factories.virtual.TimelineItemVirtualFactory
+import io.element.android.libraries.dateformatter.test.FakeDaySeparatorFormatter
 
 internal fun aTimelineItemsFactory() = TimelineItemsFactory(
     dispatchers = testCoroutineDispatchers(),
