+++ libraries/dateformatter/api/proguard-rules.pro
