+++ libraries/dateformatter/api/src/main/AndroidManifest.xml
