+++ libraries/dateformatter/api/src/main/kotlin/io/element/android/libraries/dateformatter/api/DaySeparatorFormatter.kt
@@ -14,17 +14,8 @@
  * limitations under the License.
  */
 
-package io.element.android.features.roomlist
+package io.element.android.libraries.dateformatter.api
 
-import io.element.android.libraries.dateformatter.api.LastMessageFormatter
-
-class FakeLastMessageFormatter : LastMessageFormatter {
-    private var format = ""
-    fun givenFormat(format: String) {
-        this.format = format
-    }
-
-    override fun format(timestamp: Long?): String {
-        return format
-    }
+interface DaySeparatorFormatter {
+    fun format(timestamp: Long): String
 }
