--- libraries/dateformatter/src/main/kotlin/io/element/android/libraries/dateformatter/impl/DefaultLastMessageFormatter.kt
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2023 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.element.android.libraries.dateformatter.impl
-
-import com.squareup.anvil.annotations.ContributesBinding
-import io.element.android.libraries.dateformatter.LastMessageFormatter
-import io.element.android.libraries.di.AppScope
-import javax.inject.Inject
-
-@ContributesBinding(AppScope::class)
-class DefaultLastMessageFormatter @Inject constructor(
-    private val localDateTimeProvider: LocalDateTimeProvider,
-    private val dateFormatters: DateFormatters,
-) : LastMessageFormatter {
-
-    override fun format(timestamp: Long?): String {
-        if (timestamp == null) return ""
-        val currentDate = localDateTimeProvider.providesNow()
-        val dateToFormat = localDateTimeProvider.providesFromTimestamp(timestamp)
-        val isSameDay = currentDate.date == dateToFormat.date
-        return when {
-            isSameDay -> {
-                dateFormatters.formatTime(dateToFormat)
-            }
-            else -> {
-                dateFormatters.formatDate(
-                    dateToFormat = dateToFormat,
-                    currentDate = currentDate,
-                    useRelative = true
-                )
-            }
-        }
-    }
-}
