--- libraries/dateformatter/src/test/kotlin/io/element/android/libraries/dateformatter/impl/FakeClock.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright (c) 2023 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.element.android.libraries.dateformatter.impl
-
-import kotlinx.datetime.Clock
-import kotlinx.datetime.Instant
-
-class FakeClock : Clock {
-    private var instant: Instant = Instant.fromEpochMilliseconds(0)
-
-    fun givenInstant(instant: Instant) {
-        this.instant = instant
-    }
-
-    override fun now(): Instant = instant
-}
