+++ plugins/src/main/kotlin/extension/DependencyHandleScope.kt
@@ -55,7 +55,7 @@ fun DependencyHandlerScope.allLibraries() {
     implementation(project(":libraries:matrixui"))
     implementation(project(":libraries:core"))
     implementation(project(":libraries:architecture"))
-    implementation(project(":libraries:dateformatter"))
+    implementation(project(":libraries:dateformatter:api"))
     implementation(project(":libraries:di"))
 }
