+++ samples/minimal/build.gradle.kts
@@ -52,7 +52,8 @@ dependencies {
     implementation(projects.libraries.designsystem)
     implementation(projects.libraries.architecture)
     implementation(projects.libraries.core)
-    implementation(projects.libraries.dateformatter)
+    implementation(projects.libraries.dateformatter.api)
+    implementation(projects.libraries.dateformatter.impl)
     implementation(projects.features.roomlist)
     implementation(projects.features.login)
     implementation(libs.coroutines.core)
