+++ settings.gradle.kts
@@ -44,7 +44,9 @@ include(":libraries:matrix:impl")
 include(":libraries:matrix:test")
 include(":libraries:matrixui")
 include(":libraries:textcomposer")
-include(":libraries:dateformatter")
+include(":libraries:dateformatter:api")
+include(":libraries:dateformatter:impl")
+include(":libraries:dateformatter:test")
 include(":libraries:elementresources")
 include(":libraries:ui-strings")
 include(":libraries:testtags")
